search function, single post template, all posts render below homepage, rename folder 'switchDisplay' to 'views'
