+++ client/src/components/header/login/authentication-button.tsx
@@ -6,9 +6,6 @@ import { useAuth0 } from '@auth0/auth0-react';
 
 const AuthenticationButton = () => {
   const { isAuthenticated } = useAuth0();
-  useEffect(() => {
-  }, [isAuthenticated])
-
 
   return isAuthenticated ? <LogoutButton /> : <LoginButton />;
 };
