+++ client/src/components/main/Main.tsx
@@ -10,7 +10,7 @@ export default function Main(props:any) {
             {/* toggle className 'homepage'  on h1 and h2 elements for  homepage display*/}
             <Link to='/' >
                 <h1>CRYPT-ID</h1></Link >
-            <h2 className={ home ? 'tagline homepage' : 'tagline'}>a place for believers, theorists, and those who have seen the unexplainable....</h2>
+                <h2 className={ home ? 'tagline homepage' : 'tagline'}>a place for believers, theorists, and those who have seen the unexplainable....</h2>
             <Outlet />
         </main>
     )
