+++ client/src/components/views/homepage/homepage.scss
