+++ client/src/components/views/searchLanding/searchLanding.scss
