+++ client/src/components/views/singlePost/post.scss
