+++ client/src/components/views/singlePost/post.tsx
